requirements.txt
flask 
python-dotenv
openai
aiohttp
requests
google-generativeai
pillow